Extract Puma::Cluster::WorkerHandle to a separate file
